[PRMT-TECH] ignored for a month for now false positive vulnerabilites depending on libpq
The vulnerability, yet to be addressed, resides in the external libpq C library,
used to interact vith Postgres.
The risk seems a possible denial of service, if the dependency is invoked wrongly.

We don't interact with libpq directly but via other library, and if we were vulnerable
to it we should experience a denial of service quite frankly. Therefore, I see the result
of npm audit as a false positive. I set it to be ignored for a month, if in a month this
hasn't been addressed yet I'd suggest to ignore this for loger or permanently.

For more details:

https://github.com/advisories/GHSA-j32j-2hxv-rqf7
https://nvd.nist.gov/vuln/detail/CVE-2022-25852
https://security.snyk.io/vuln/SNYK-JS-LIBPQ-2392366